refactor(api): créer une factory de contrôleur pour deleteComment refers to https://trello.com/c/evsOelCU/2717
Les contrôleurs actionComment.deleteComment et shantytownComment.moderateComment
(l'équivalent "delete" côté site) étaient strictement identiques à
l'exception du service appelé. Extraction dans makeDeleteCommentController,
une factory générique paramétrée par la fonction de service, réutilisée par
les deux contrôleurs. Comportement HTTP inchangé (mêmes codes ERROR_RESPONSES,
même statut 200, même forme de réponse).
