+++ packages/api/server/controllers/shantytownComment/moderateComment/shantytownComment.moderateComment.ts
@@ -1,29 +1,4 @@
-
 import shantytownCommentService from '#server/services/shantytownComment';
-import { ShantytownEnrichedComment } from '#root/types/resources/ShantytownCommentEnriched.d';
+import makeDeleteCommentController from '#server/controllers/_common/makeDeleteCommentController';
 
-const { deleteComment } = shantytownCommentService;
-
-const ERROR_RESPONSES = {
-    fetch_failed: { code: 400, message: 'Une lecture en base de données a échoué' },
-    data_incomplete: { code: 400, message: 'Données manquantes' },
-    permission_denied: { code: 403, message: 'Permission refusée' },
-    delete_failed: { code: 500, message: 'Une suppression en base de données a échoué' },
-    undefined: { code: 500, message: 'Une erreur inconnue est survenue' },
-};
-
-
-export default async function moderateComment(req, res, next) {
-    let comments: { comments: ShantytownEnrichedComment[] };
-    try {
-        comments = await deleteComment(req.user, Number.parseInt(req.params.id, 10), Number.parseInt(req.params.commentId, 10), req.body.message);
-    } catch (error) {
-        const { code, message }: { code: number; message: string } = ERROR_RESPONSES[error?.code] ?? ERROR_RESPONSES.undefined;
-        res.status(code).send({
-            user_message: message,
-        });
-        return next(error.nativeError ?? error);
-    }
-
-    return res.status(200).send(comments);
-}
+export default makeDeleteCommentController(shantytownCommentService.deleteComment);
